Move the List command code to a concern
The list commands are not well suited to inheritance as it forms to
rigid of a structure.

Instead the code has been moved to a concern as a composition adds
greater flexibility
